fix _GLOBAL_TIMEOUT type to int

+++ python3.6/run/runtime-mock.py
@@ -27,7 +27,7 @@ _GLOBAL_EVENT_BODY = sys.argv[2] if len(sys.argv) > 2 else os.environ.get('AWS_L
 _GLOBAL_FCT_NAME = os.environ.get('AWS_LAMBDA_FUNCTION_NAME', 'test')
 _GLOBAL_VERSION = os.environ.get('AWS_LAMBDA_FUNCTION_VERSION', '$LATEST')
 _GLOBAL_MEM_SIZE = os.environ.get('AWS_LAMBDA_FUNCTION_MEMORY_SIZE', 1536)
-_GLOBAL_TIMEOUT = os.environ.get('AWS_LAMBDA_FUNCTION_TIMEOUT', 300)
+_GLOBAL_TIMEOUT = int(os.environ.get('AWS_LAMBDA_FUNCTION_TIMEOUT', 300))
 _GLOBAL_REGION = os.environ.get('AWS_REGION', 'us-east-1')
 _GLOBAL_ACCOUNT_ID = os.environ.get('AWS_ACCOUNT_ID', _random_account_id())
 _GLOBAL_ACCESS_KEY_ID = os.environ.get('AWS_ACCESS_KEY_ID', 'SOME_ACCESS_KEY_ID')
@@ -139,7 +139,7 @@ def report_done(invokeid, errortype, result):
         eprint("END RequestId: %s" % invokeid)
 
         duration = int((time.time() - _GLOBAL_START_TIME) * 1000)
-        billed_duration = min(100 * ((duration / 100) + 1), int(_GLOBAL_TIMEOUT) * 1000)
+        billed_duration = min(100 * ((duration / 100) + 1), _GLOBAL_TIMEOUT * 1000)
         max_mem = int(resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / 1024)
 
         eprint(
@@ -164,7 +164,7 @@ def log_sb(msg):
     return
 
 def get_remaining_time():
-    return ((int(_GLOBAL_TIMEOUT) * 1000) - int((time.time() - _GLOBAL_START_TIME) * 1000))
+    return ((_GLOBAL_TIMEOUT * 1000) - int((time.time() - _GLOBAL_START_TIME) * 1000))
 
 def send_console_message(msg):
     eprint(msg)
